feat: significant refactor of the codebase
Split the crate into multiple modules to make it easier to naviagte and maintain,
in particular isolated the WAFObject into an object module; the log facility into
a log module.

Configured stringent clippy configuration (this most likely will be loosened
as we see lints that do not make sense to us in the future).

The crate no longer has a dynamic_lib feature, now it has a static feature
instead (not enabled by default, although we can reconsider).