feat: significant refactor of the codebase (#5)
Split the crate into multiple modules to make it easier to navigate and
maintain,
in particular isolated the `WAFObject` (and related) into an `object`
module; the log facility into a `log` module, etc...

Configured stringent clippy configuration (this most likely will be
loosened
as we see lints that do not make sense to us in the future).

The crate no longer has a `dynamic_lib` or `static_lib` feature, as it
turns out `libddwaf` bundles symbols that interfere with rust's panic
unwinding (causing it to fail to initiate a panic).

Moved a bunch of tests under `tests`, so more code is tested
crate-external.
Also improved test coverage a bit.

This also adds CI set-up.

---

Jira: APPSEC-58070