feat: (re)-introduce `static` feature (#11)
Removing the `Unwind*` objects from the `libddwaf.a` file resolves
conflicts between the rust standard library unwinder and the LLVM
unwinder that is baked into `libddwaf`. This allows the `static` feature
to be enabled by default (disabling it still allows for dynamically
linking `libddwaf` if needed).

---

Also:
- un-does the attempt to use the `C-unwind` ABI (which is not needed
here),
- introduces a matrix test across multiple platforms,
- migrates to the Datadog-standard `LICENSES-3rdparty.csv` format using
the blessed tool,
- improves `README.md` documentation a little